We now persist KobwebModels across restarts
We also show an informative pop-up to users to kickstart
the process if we detect they haven't generated the
persisted data yet.

The flow at this point is:

* The Kobweb plugin looks for persisted data
* If it can't find it, it asks the user to sync
* A sync, when done, installs KobwebModels as before.
  But now, those models are also now persisted.
* After a restart, the Kobweb plugin will detect
  the cached data and use it to repopulate KobwebModels,
  no expensive syncing required.

There's also logic which checks if the project has at
least one Kobweb dependency. If not, we assume we're
in a non-Kobweb project where the user just happens to
have the Kobweb plugin installed for a different project.
In this case, we make sure we don't show users the pop-up
asking them to sync.