Show a "You should sync" message if necessary
A full Gradle sync is required for us to have up-to-date
Kobweb metadata information. However, if a user just
installed the plugin, they won't have that information and
won't realize it. So we show them a notification to let
them know.

The user shouldn't see this popup if...

* they don't depend on Kobweb anywhere throughout the project
* they already did an import through some other means (i.e. if
  you open a new Kobweb project from scratch, it kicks off a
  scan that way)
* they finished a scan in a previous IDE session

We also use this information to avoid caching Kobweb state /
processing Kobweb functionality if the plugin is considered
disabled for the current project.